+++ galois/_codes/_cyclic.py
@@ -5,9 +5,8 @@ from typing import Optional
 
 import numpy as np
 
-from .._fields import FieldArray
+from .._fields import FieldArray, Poly
 from .._overrides import set_module
-from .._polys import Poly
 
 __all__ = ["poly_to_generator_matrix", "roots_to_parity_check_matrix"]
