--- galois/_polys/__init__.py
@@ -1,9 +1,0 @@
-"""
-A subpackage containing polynomials over Galois fields.
-"""
-# pylint: disable=redefined-builtin
-
-from ._factor import *
-from ._functions import *
-from ._math import *
-from ._poly import *
