+++ setup.cfg
@@ -42,7 +42,7 @@ project_urls =
     Twitter = https://twitter.com/galois_py
 
 [options]
-packages = galois, galois._codes, galois._databases, galois._fields, galois._polys
+packages = galois, galois._codes, galois._databases, galois._fields
 include_package_data = true
 test_suite = tests
 python_requires = >=3.6
